pyproject.toml
[project]
name = "better-nested-serializer"
version = "0.1.2"
description = "A better DRF [Django REST Framework] serializer that normalizes nested objects into separate lists for improved performance"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Rudra", email = "rudra.ofive@gmail.com"},
]
keywords = ["django", "rest-framework", "serializer", "performance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "deepdiff>=8.6.1",
    "deepmerge>=2.0",
    "djangorestframework>=3",
]

[project.urls]
Homepage = "https://github.com/sree93/better_drf_nested_serializer"
Repository = "https://github.com/sree93/better_drf_nested_serializer"
Issues = "https://github.com/sree93/better_drf_nested_serializer/issues"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Include single-file top-level modules in the distribution
py-modules = ["helpers"]

[tool.setuptools.packages.find]
# Find import packages (those with __init__.py)
where = ["."]
include = [
    "exceptions*",
    "serializers*",
]
exclude = [
    "test_app",
    "test_app.*",
    "**/migrations*",
    "**/test*",
    "**/*test*",
    "__pycache__*",
    ".git*",
    ".github*",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "build>=1.3.0",
    "deepdiff>=8.6.1",
    "twine>=6.2.0",
]
